+++ Assets/Scripts/solid/EnemyStateInfoPanel.cs
@@ -100,11 +100,11 @@ public class EnemyStateInfoPanel : MonoBehaviour
     {
         
         
-            happinesText.text = "Happines: " + currnetState.Morele.ToString();
-            totalArmyPowerText.text = "Army: " + currnetState.TotalArmyPower.ToString();
-            mainResTypeText.text = "Main Resoruce: " + ((MainResourceType)currnetState.Resources).ToString();
+            happinesText.text = "Happines: " + ((int)currnetState.Morele).ToString();
+        totalArmyPowerText.text = "Army: " + ((int)currnetState.TotalArmyPower).ToString();
+        mainResTypeText.text = "Main Resoruce: " + ((MainResourceType)currnetState.Resources).ToString();
+       
 
-        
     }
   
     public void FailedEspionage()
