+++ Assets/Scripts/solid/NaturalStateInfoPanel.cs
@@ -55,8 +55,8 @@ public class NaturalStateInfoPanel : MonoBehaviour
     {
 
         {
-            happinesText.text = "Happines: " + currnetState.Morele.ToString();
-            totalArmyPowerText.text = "Army: " + currnetState.TotalArmyPower.ToString();
+            happinesText.text = "Happines: " +( (int)currnetState.Morele).ToString();
+            totalArmyPowerText.text = "Army: " + ((int)currnetState.TotalArmyPower).ToString();
             mainResTypeText.text = "Main Resoruce" + ((MainResourceType)currnetState.Resources).ToString();
 
         }
